Upgrades dependencies
The upgrade of Typescript and Jest had some follow-ups that needed
fixing.

The typescript strict-check now requires work to be done in catch
blocks where the thrown object is automatically of type `unknown`.

The `tsConfig` setting is deprecated and should be `tsconfig`
now. Also the `jsdom` package must be added explicitly.