WB2-183 Remove integration with Scatter.js
The Wombat UAL library uses Scatter.js to communicate with
Wombat. This is perfectly viable as Wombat exposes the Scatter API,
but brings in a lot of dependencies. This can result in
trouble (*). Additionally it bloats up our package even though we are
really not using that much of the library anyway.

The integration is completely removed, now only `window.scatter` as
provided by Wombat is used.

Note: This means when we develop the Wombat desktop application which
probably will work via Websockets (but definitely not via
`window.scatter` as that's technically impossible), additional work
will need to be done.

(*) E.g. the transitive dependency `device-uuid` of Scatter.js does
not build in production mode in dapps using `nuxt`. This is due to a
bug in the nuxt build system dependencies, but it's an unfortunate
situation for us anyway.

The babel runtime package seems not to be needed anymore - I think it
was connected to the jest mock. Not sure what that is about.