Refactoring & minor changes
Adds an error state to the authenticator
========================================

I'm not sure if any clients really use it, but since it's kind of
cheap to add I don't see anything against it.

The error is set either when Wombat is not present or connecting to it
fails.

Use new Wombat domain
=====================

Uses the new www.wombat.app domain instead of genialwombat.io

Small changes
=============

* Makes the appName property `readonly` as it's not required to be
changed
* Correctly resets state in `reset` and logs errors that happen when
"re-calling" `init`
* Removes a documentation that is not true anymore
* Removes linebreaks that I find unnecessary
* Renames one caught object to `e` to follow the overall style
* Removes the unused `isMobile` method